Major reorganising and addition of 3D plot
I did some major reorganising for the code, which includes (but is not limited to) moving more functionalities to the utils.py file, creating a separate main.py file, changed the style of the docstrings.

I also added the ability to visualise not only R^2 but also R^3. Both use the same function. I am not sure if that is the best solution.